use common code
